sign_language_model/train_classifier.py
import pickle

import numpy as np
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score
from sklearn.model_selection import train_test_split

data_dict = pickle.load(open("data.pickle", "rb"))

data = np.asarray(data_dict["data"])
labels = np.asarray(data_dict["labels"])
x_train, x_test, y_train, y_test = train_test_split(
    data, labels, test_size=0.2, train_size=0.8, shuffle=True, stratify=labels
)

model = RandomForestClassifier()

model.fit(x_train, y_train)

y_predict = model.predict(x_test)

score = accuracy_score(y_predict, y_test)

print("{:.2f}% of samples were correctly classified!".format(score * 100))

f = open("model.p", "wb")
pickle.dump({"model": model}, f)
f.close()
