requirements.txt
# Temel Kütüphaneler
numpy>=1.19.0
opencv-python>=4.5.0
mediapipe>=0.8.9
PyYAML>=6.0
python-dotenv>=0.19.0

# Makine Öğrenmesi
scikit-learn==1.2.0

# Çeviri İşlemleri
googletrans==3.1.0a0

# Görüntü İşleme ve Çizim
matplotlib>=3.7.0

# Veri İşleme ve Serileştirme
pickle-mixin>=1.0.2

# GUI Bileşenleri
customtkinter>=5.2.0

# Test Araçları
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-html>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0
coverage>=7.2.0

# Not: Aşağıdaki kütüphaneler Python ile birlikte gelir:
# - tkinter: GUI için
# - winsound: Ses işleme için (Windows)
# - logging: Loglama için
# - unittest: Birim testler için 